Stop parsers from knowing about sessions.
It doesn't make a whole lot of sense that the thing which will be
responsible for telling you to restart a session should know what
the session is.

Move things into a CardCommand class for now. (As opposed to a
command to restart the game or to flip the stack.) Will worry more
about how to realize commands in the future.