Move stacks of cards!!!
Note that as an artifact of implementation choices, this design
has a fair amount of validation built in. In particular, you can't
move hidden cards, cards in the stack, cards that are buried, or
cards in a discard pile.

I'm taking this as a small sign that I'm doing something right.